Move sub-functions into views
